requirements-dev.txt
pytest-cov==2.8.1
pytest-flask==0.15.1
pytest-random-order==1.0.4
pytest==5.3.5
